Generate property count constraint for map members
#84
